==> 0-Formula1-2-data_merge-historical-race.ipynb <==
# Libraries
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

"""
# Data Access
car_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/car_data_2024.csv")
circuit_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/circuit_2024.csv")
control_message_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/control_message_2024.csv")
lap_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/lap_2024.csv", low_memory=False)
position_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/position_2024.csv")
result_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/result_2024.csv")
session_status_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/session_status_2024.csv")
track_status_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/track_status_2024.csv")
weather_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/weather_2024.csv")
"""

# Race Data Access
race_car_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/car_data_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_circuit_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/circuit_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_control_message_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/control_message_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_lap_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/lap_mexico_grand_prix_2018-2019-2021-2022-2023.csv", low_memory=False)
race_position_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/position_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_result_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/result_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_session_status_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/session_status_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_track_status_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/track_status_mexico_grand_prix_2018-2019-2021-2022-2023.csv")
race_weather_data = pd.read_csv("C:/Users/yunus/Downloads/Mexican GP Data Challenge Datasets/ocean_mexican_gp_data_challenge_datasets/mexican_gp_race_datasets_2018-2019-2021-2022-2023/weather_mexico_grand_prix_2018-2019-2021-2022-2023.csv")


# Data Merge
# 4. lap_data
# Extract days and hours
race_lap_data['Days'] = race_lap_data['Time'].apply(lambda x: int(x.split(' ')[0]))
race_lap_data['Hours'] = race_lap_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_lap_data['Minutes'] = race_lap_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_lap_data)

# Extract days and hours
race_weather_data['Days'] = race_weather_data['Time'].apply(lambda x: int(x.split(' ')[0]))
race_weather_data['Hours'] = race_weather_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_weather_data['Minutes'] = race_weather_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_weather_data)

# Remove duplicates in weather_data based on 'eventname' and 'hours'
race_weather_data1 = race_weather_data.drop_duplicates(subset=['Year', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data1 = pd.merge(race_lap_data, race_weather_data1, on=['Year', 'Hours', 'Minutes'], how='left')

# The number of rows in merged_df should now match df1

print(race_lap_data1)

race_lap_data1.head(), race_lap_data1.info()

# car_data
# Extract days and hours
race_car_data['Days'] = race_car_data['SessionTime'].apply(lambda x: int(x.split(' ')[0]))
race_car_data['Hours'] = race_car_data['SessionTime'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_car_data['Minutes'] = race_car_data['SessionTime'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_car_data)

race_car_data.head(), race_car_data.info(), race_car_data.describe()

race_car_data.describe(include='object')

# Remove duplicates in weather_data based on 'eventname' and 'hours'
race_car_data1 = race_car_data.drop_duplicates(subset=['Year', 'DriverName', 'Hours', 'Minutes'])


race_car_data1.rename(columns={'DriverName': 'Driver'}, inplace=True)


race_car_data1.head(), race_car_data1.info(), race_car_data1.describe(), race_car_data1.describe(include='object')

# Remove duplicates in weather_data based on 'eventname' and 'hours'
# car_data1 = car_data.drop_duplicates(subset=['EventName', 'DriverName', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data2 = pd.merge(race_lap_data1, race_car_data1, on=['Year', 'Driver', 'Hours', 'Minutes'], how='left')

# The number of rows in merged_df should now match df1

print(race_lap_data2)

race_lap_data2.head(), race_lap_data2.info()

# Data Merge
# 5. position
race_position_data.info()

race_position_data.head(), race_position_data.tail()

race_position_data.rename(columns={'DriverName': 'Driver'}, inplace=True)

# position_data
# Extract days and hours
race_position_data['Days'] = race_position_data['Time'].apply(lambda x: int(x.split(' ')[0]))
race_position_data['Hours'] = race_position_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_position_data['Minutes'] = race_position_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_position_data)

# Remove duplicates in position_data based on 'eventname' and 'hours'
race_position_data1 = race_position_data.drop_duplicates(subset=['Year', 'Driver', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data3 = pd.merge(race_lap_data2, race_position_data1, on=['Year', 'Driver', 'Hours', 'Minutes'], how='left', suffixes=('_left', '_right'))

# The number of rows in merged_df should now match df1

print(race_lap_data3)

race_lap_data3.head(), race_lap_data3.info()

# Data Merge
# 6. result_data

race_result_data.info(), race_result_data.head()

race_result_data.rename(columns={'Abbreviation': 'Driver'}, inplace=True)

# Remove duplicates in result_data based on 'eventname' and 'hours'
# result_data1 = result_data.drop_duplicates(subset=['EventName', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data4 = pd.merge(race_lap_data3, race_result_data, on=['Year', 'Driver'], how='left', suffixes=('_left', '_right'))

# The number of rows in merged_df should now match df1

print(race_lap_data4)

race_lap_data4.info(), race_lap_data4.describe(), race_lap_data4.describe(include='object')

race_lap_data4['Date_left']

race_lap_data4['Date_right']

# Data Merge
# 3. control_message (Cancelled due to similar data from track_status)
# print(control_message_data)

# Data Merge
# 7. track_status
print(race_track_status_data)

race_track_status_data.head(20)

# track_status_data
# Extract days and hours
race_track_status_data['Days'] = race_track_status_data['Time'].apply(lambda x: int(x.split(' ')[0]))
race_track_status_data['Hours'] = race_track_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_track_status_data['Minutes'] = race_track_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_track_status_data)

# Remove duplicates in track_status_data based on 'eventname' and 'hours'
race_track_status_data1 = race_track_status_data.drop_duplicates(subset=['Year', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data5 = pd.merge(race_lap_data4, race_track_status_data1, on=['Year', 'Hours', 'Minutes'], how='left', suffixes=('_left2', '_right2'))

# The number of rows in merged_df should now match df1

print(race_lap_data5)

race_lap_data5.info()

# Data Merge
# 7. session_status_data
print(race_session_status_data)

# session_status_data
# Extract days and hours
race_session_status_data['Days'] = race_session_status_data['Time'].apply(lambda x: int(x.split(' ')[0]))
race_session_status_data['Hours'] = race_session_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
race_session_status_data['Minutes'] = race_session_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(race_session_status_data)

# Remove duplicates in track_status_data based on 'eventname' and 'hours'
race_session_status_data1 = race_session_status_data.drop_duplicates(subset=['Year', 'Hours', 'Minutes'])

# Now perform the left join
race_lap_data6 = pd.merge(race_lap_data5, race_session_status_data1, on=['Year', 'Hours', 'Minutes'], how='left', suffixes=('_left3', '_right3'))

# The number of rows in merged_df should now match df1

print(race_lap_data6)

race_lap_data6.info(), race_lap_data6.head()

# Save DataFrame to CSV
race_lap_data6.to_csv('race_all_data.csv', index=False)

"""
# 2. circuit_data
# Extract days and hours
circuit_data['Days'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[0]))
circuit_data['Hours'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
circuit_data['Minutes'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(circuit_data)
"""

"""
# 3. control_message_data
# Extract days and hours
control_message_data['Days'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[0]))
control_message_data['Hours'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
control_message_data['Minutes'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(control_message_data)
"""
