==> 1-Formula1-2-data_merge-2024-race.ipynb <==
# Libraries
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

# Data Access
car_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/car_data_2024.csv")
circuit_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/circuit_2024.csv")
control_message_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/control_message_2024.csv")
lap_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/lap_2024.csv", low_memory=False)
position_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/position_2024.csv")
result_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/result_2024.csv")
session_status_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/session_status_2024.csv")
track_status_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/track_status_2024.csv")
weather_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/weather_2024.csv")


# Data Merge
# 4. Lap_data
# lap_data = pd.read_csv("C:/Users/yunus/Downloads/Dataset-f1/ocean_data_challenge_F1_data/races_2024/lap_2024.csv", low_memory=False)

lap_data.head()

# lap_data
# Extract days and hours
lap_data['Days'] = lap_data['Time'].apply(lambda x: int(x.split(' ')[0]))
lap_data['Hours'] = lap_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
lap_data['Minutes'] = lap_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(lap_data)

# Extract days and hours
weather_data['Days'] = weather_data['Time'].apply(lambda x: int(x.split(' ')[0]))
weather_data['Hours'] = weather_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
weather_data['Minutes'] = weather_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(weather_data)

# Remove duplicates in weather_data based on 'eventname' and 'hours'
weather_data1 = weather_data.drop_duplicates(subset=['EventName', 'Hours', 'Minutes'])

# Now perform the left join
lap_data1 = pd.merge(lap_data, weather_data1, on=['EventName', 'Hours', 'Minutes'], how='left')

# The number of rows in merged_df should now match df1

print(lap_data1)

lap_data1.head(), lap_data1.info()

# car_data
# Extract days and hours
car_data['Days'] = car_data['SessionTime'].apply(lambda x: int(x.split(' ')[0]))
car_data['Hours'] = car_data['SessionTime'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
car_data['Minutes'] = car_data['SessionTime'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(car_data)

car_data.head(), car_data.info(), car_data.describe()

car_data.describe(include='object')

# Remove duplicates in weather_data based on 'eventname' and 'hours'
car_data1 = car_data.drop_duplicates(subset=['EventName', 'DriverName', 'Hours', 'Minutes'])


car_data1.rename(columns={'DriverName': 'Driver'}, inplace=True)


car_data1.head(), car_data1.info(), car_data1.describe(), car_data1.describe(include='object')

# Remove duplicates in weather_data based on 'eventname' and 'hours'
# car_data1 = car_data.drop_duplicates(subset=['EventName', 'DriverName', 'Hours', 'Minutes'])

# Now perform the left join
lap_data2 = pd.merge(lap_data1, car_data1, on=['EventName', 'Driver', 'Hours', 'Minutes'], how='left')

# The number of rows in merged_df should now match df1

print(lap_data2)

lap_data2.head(), lap_data2.info()

# Data Merge
# 5. position
position_data.info()

position_data.head(), position_data.tail()

position_data.rename(columns={'DriverName': 'Driver'}, inplace=True)

# position_data
# Extract days and hours
position_data['Days'] = position_data['Time'].apply(lambda x: int(x.split(' ')[0]))
position_data['Hours'] = position_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
position_data['Minutes'] = position_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(position_data)

# Remove duplicates in position_data based on 'eventname' and 'hours'
position_data1 = position_data.drop_duplicates(subset=['EventName', 'Driver', 'Hours', 'Minutes'])

# Now perform the left join
lap_data3 = pd.merge(lap_data2, position_data1, on=['EventName', 'Driver', 'Hours', 'Minutes'], how='left', suffixes=('_left', '_right'))

# The number of rows in merged_df should now match df1

print(lap_data3)

lap_data3.head(), lap_data3.info()

# Data Merge
# 6. result_data

result_data.info(), result_data.head()

result_data.rename(columns={'Abbreviation': 'Driver'}, inplace=True)

# Remove duplicates in result_data based on 'eventname' and 'hours'
# result_data1 = result_data.drop_duplicates(subset=['EventName', 'Hours', 'Minutes'])

# Now perform the left join
lap_data4 = pd.merge(lap_data3, result_data, on=['EventName', 'Driver'], how='left', suffixes=('_left', '_right'))

# The number of rows in merged_df should now match df1

print(lap_data4)

lap_data4.info(), lap_data4.describe(), lap_data4.describe(include='object')

lap_data4['Date_left']

lap_data4['Date_right']

# Data Merge
# 3. control_message (Cancelled due to similar data from track_status)
# print(control_message_data)

# Data Merge
# 7. track_status
print(track_status_data)

track_status_data.head(20)

# track_status_data
# Extract days and hours
track_status_data['Days'] = track_status_data['Time'].apply(lambda x: int(x.split(' ')[0]))
track_status_data['Hours'] = track_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
track_status_data['Minutes'] = track_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(track_status_data)

# Remove duplicates in track_status_data based on 'eventname' and 'hours'
track_status_data1 = track_status_data.drop_duplicates(subset=['EventName', 'Hours', 'Minutes'])

# Now perform the left join
lap_data5 = pd.merge(lap_data4, track_status_data1, on=['EventName', 'Hours', 'Minutes'], how='left', suffixes=('_left2', '_right2'))

# The number of rows in merged_df should now match df1

print(lap_data5)

lap_data5.info()

# Data Merge
# 7. session_status_data
print(session_status_data)

# session_status_data
# Extract days and hours
session_status_data['Days'] = session_status_data['Time'].apply(lambda x: int(x.split(' ')[0]))
session_status_data['Hours'] = session_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
session_status_data['Minutes'] = session_status_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(session_status_data)

# Remove duplicates in track_status_data based on 'eventname' and 'hours'
session_status_data1 = session_status_data.drop_duplicates(subset=['EventName', 'Hours', 'Minutes'])

# Now perform the left join
lap_data6 = pd.merge(lap_data5, session_status_data1, on=['EventName', 'Hours', 'Minutes'], how='left', suffixes=('_left3', '_right3'))

# The number of rows in merged_df should now match df1

print(lap_data6)

lap_data6.info(), lap_data6.head()

# Save DataFrame to CSV
lap_data6.to_csv('all_data.csv', index=False)

"""
# 2. circuit_data
# Extract days and hours
circuit_data['Days'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[0]))
circuit_data['Hours'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
circuit_data['Minutes'] = circuit_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(circuit_data)
"""

"""
# 3. control_message_data
# Extract days and hours
control_message_data['Days'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[0]))
control_message_data['Hours'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[0]))
control_message_data['Minutes'] = control_message_data['Time'].apply(lambda x: int(x.split(' ')[2].split(':')[1]))

# View the updated DataFrame
print(control_message_data)
"""
