==> 2-Formula1-2-data_merge-all.ipynb <==
# Libraries
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

race_all_data = pd.read_csv("C:/Users/yunus/Downloads/race_all_data.csv")
all_data = pd.read_csv("C:/Users/yunus/Downloads/all_data.csv", low_memory=False)

race_all_data.info(), all_data.info()

race_all_data.insert(loc=0, column='EventName', value='Mexican Grand Prix')

all_data.insert(loc=1, column='Year', value=2024)

# race_all_data['EventName'] = 'Mexican Grand Prix'

# all_data['Year'] = 2024

race_all_data.info(), all_data.info()

race_all_data.head(), all_data.head()

merged_data = pd.concat([all_data, race_all_data], ignore_index=True)

merged_data.info()

# Save DataFrame to CSV
merged_data.to_csv('merged_data.csv', index=False)

# Save DataFrame to Excel
merged_data.to_excel('merged_data.xlsx', index=False)
