==> 5-Formula1-2-final-report-codes.ipynb <==
# Libraries
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
from sklearn.preprocessing import StandardScaler

# Data Access
merged_data = pd.read_csv("C:/Users/yunus/Downloads/Formula1-2/output.csv", low_memory=False)


# Checking data summary
merged_data.info()

# Data Preprocessing
# 1. Handling Missing Values
# Drop columns with very high missingness and repeated data due to merge
merged_data.drop(['Deleted', 'FastF1Generated', 'IsAccurate', 'Message', 'IsPersonalBest', 'Q1', 'Q2', 'Q3','PitOutTime', 'PitInTime', 
                  'DeletedReason', 'HeadshotUrl', 'Time_right2', 'Status_left3', 'Days_left3', 'Time', 'Status_right3', 'Days_right3',
                 'Time_y', 'Date_left', 'Source_left', 'Time_left', 'SessionTime_left', 'Days_left', 'Date_right', 'Status_left', 
                  'Source_right', 'Time_right', 'SessionTime_right', 'Days_right', 'DriverNumber_right', 'BroadcastName', 'DriverId', 
                  'TeamName', 'TeamColor', 'TeamId', 'FirstName', 'LastName', 'FullName', 'CountryCode', 'Time_left2', 'Status_right', 
                   'LapStartDate', 'Days_x', 'Days_y'],
                 axis=1, inplace=True)


# Save the model ready dataframe to CSV
merged_data.to_csv('modeldata.csv', index=False)

# ML MODELS
from catboost import CatBoostRegressor, CatBoostClassifier
from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, accuracy_score, r2_score
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor, DecisionTreeClassifier
import xgboost as xgb
from xgboost import XGBRegressor, XGBClassifier
import lightgbm as lgb
from lightgbm import LGBMClassifier

# Checking data summary
merged_data.info()

# Selecting relevant columns for modeling
target_columns = ['Number_of_Laps', 'Used_Tire_Compound', 'Avg_Lap_Time', 'Number_of_Stints']

feature_columns = ['EventName', 'Driver', 'GridPosition', 'Year', 'TyreLife', 'FreshTyre', 'TrackStatus', 'Rainfall',
                   'Speed', 'Throttle', 'Brake', 'DRS', 'ClassifiedPosition', 'AirTemp', 'Humidity', 'Pressure', 'TrackTemp', 'WindSpeed']

# Cleaning data: dropping rows with missing target values and converting 'Avg_Lap_Time' to seconds
data_cleaned = merged_data.dropna()
data_cleaned['Avg_Lap_Time'] = pd.to_timedelta(data_cleaned['Avg_Lap_Time']).dt.total_seconds()

# Splitting data for each target
X = data_cleaned[feature_columns]

# Convert categorical columns to category dtype
X['Driver'] = X['Driver'].astype('category')
X['EventName'] = X['EventName'].astype('category')
X['Year'] = X['Year'].astype('category')
X['ClassifiedPosition'] = X['ClassifiedPosition'].astype('category')
X['FreshTyre'] = X['FreshTyre'].astype('category')
X['Brake'] = X['Brake'].astype('category')
X['Rainfall'] = X['Rainfall'].astype('category')
X['TrackStatus'] = X['TrackStatus'].astype('category')
X['DRS'] = X['DRS'].astype('category')

# 'Driver', 'EventName', 'Year', 'ClassifiedPosition', 'FreshTyre', 'Brake', 'Rainfall', 'TrackStatus', 'DRS'

# Splitting the data for each model
y_laps = data_cleaned['Number_of_Laps']
y_tires = data_cleaned['Used_Tire_Compound']
y_avg_lap_time = data_cleaned['Avg_Lap_Time']
y_stints = data_cleaned['Number_of_Stints']

X_train_laps, X_test_laps, y_train_laps, y_test_laps = train_test_split(X, y_laps, test_size=0.2, random_state=42)
X_train_tires, X_test_tires, y_train_tires, y_test_tires = train_test_split(X, y_tires, test_size=0.2, random_state=42)
X_train_avg_lap_time, X_test_avg_lap_time, y_train_avg_lap_time, y_test_avg_lap_time = train_test_split(X, y_avg_lap_time, test_size=0.2, random_state=42)
X_train_stints, X_test_stints, y_train_stints, y_test_stints = train_test_split(X, y_stints, test_size=0.2, random_state=42)

# Checking data summary
data_cleaned.info()

cat_cols = ['Driver', 'EventName', 'Year', 'ClassifiedPosition', 'FreshTyre', 'Brake', 'Rainfall', 'TrackStatus', 'DRS']

# Model 1: Predicting Number_of_Stints (Regression) - XGBoost
model_stints_xgb = XGBRegressor(n_estimators= 3500,learning_rate= 0.05,max_depth= 5, objective='reg:squarederror', 
                              eval_metric= 'rmse', enable_categorical=True, random_state=42)
model_stints_xgb.fit(X_train_stints, y_train_stints)
y_pred_stints_xgb = model_stints_xgb.predict(X_test_stints)

stints_mse_xgb = mean_squared_error(y_test_stints, y_pred_stints_xgb)
stints_mse_xgb

# Model 1.2: Predicting Number_of_Stints (Regression) - LightGBM
model_stints_lgbm = lgb.LGBMRegressor(n_estimators= 3500,learning_rate= 0.08,max_depth= 9, objective='regression', 
                              random_state=42, verbose=-1)
model_stints_lgbm.fit(X_train_stints, y_train_stints)
y_pred_stints_lgbm = model_stints_lgbm.predict(X_test_stints)

stints_mse_lgbm = mean_squared_error(y_test_stints, y_pred_stints_lgbm)
stints_mse_lgbm

# Model 1.3: Predicting Number_of_Stints (Regression) - CatBoost
model_stints_catboost = CatBoostRegressor(iterations= 3500,learning_rate= 0.05,depth= 5,objective='RMSE',
            eval_metric= 'RMSE',verbose= 500,border_count=254,task_type='CPU',use_best_model=False)
model_stints_catboost.fit(X_train_stints, y_train_stints, cat_features = cat_cols)
y_pred_stints_catboost = model_stints_catboost.predict(X_test_stints)

stints_mse_catboost = mean_squared_error(y_test_stints, y_pred_stints_catboost)
stints_mse_catboost

# Model 2: Predicting Used_Tire_Compound (Classification) - LightGBM

# Define and train the LightGBM classifier
model_tires_lgbm = LGBMClassifier(n_estimators= 3500,learning_rate= 0.08,max_depth= 5, objective='multiclass',random_state=42, 
                                  verbose=-1) # verbose=-1 disable warnings
model_tires_lgbm.fit(X_train_tires, y_train_tires)

# Make predictions
y_pred_tires_lgbm = model_tires_lgbm.predict(X_test_tires)

tires_accuracy_lgbm = accuracy_score(y_test_tires, y_pred_tires_lgbm)
tires_accuracy_lgbm

# Model 2.2: Predicting Used_Tire_Compound (Classification) - CatBoost

# Define and train the CatBoost classifier
model_tires_catboost = CatBoostClassifier(iterations= 3500,learning_rate= 0.08,depth= 9,objective='MultiClass',
            eval_metric= 'MultiClass',verbose= 500,border_count=254,task_type='CPU',use_best_model=False)

#iterations=1000, random_state=42, verbose=100
model_tires_catboost.fit(X_train_tires, y_train_tires, cat_features=cat_cols)
y_pred_tires_catboost = model_tires_catboost.predict(X_test_tires)

tires_accuracy_catboost = accuracy_score(y_test_tires, y_pred_tires_catboost)
tires_accuracy_catboost

# Model 3: Predicting Number_of_Laps (Regression) - XGBoost
model_laps_xgb = XGBRegressor(n_estimators= 3500,learning_rate= 0.05,max_depth= 5, objective='reg:squarederror', 
                              eval_metric= 'rmse', enable_categorical=True, random_state=42)
# n_estimators= 3500,learning_rate= 0.05,max_depth= 5, objective='reg:squarederror', eval_metric= 'rmse', enable_categorical=True, random_state=42
model_laps_xgb.fit(X_train_laps, y_train_laps)
y_pred_laps_xgb = model_laps_xgb.predict(X_test_laps)

laps_mse_xgb = mean_squared_error(y_test_laps, y_pred_laps_xgb)
laps_mse_xgb

# Model 3.2: Predicting Number_of_Laps (Regression) - LightGBM
model_laps_lgbm = lgb.LGBMRegressor(n_estimators= 3500,learning_rate= 0.07,max_depth= 9, objective='regression', metric='rmse',
                              random_state=42, verbose=-1, num_leaves=45, subsample=0.8, colsample_bytree=0.9)
model_laps_lgbm.fit(X_train_laps, y_train_laps)
y_pred_laps_lgbm = model_laps_lgbm.predict(X_test_laps)

laps_mse_lgbm = mean_squared_error(y_test_laps, y_pred_laps_lgbm)
laps_mse_lgbm

# Model 3.3: Predicting Number_of_Laps (Regression) - CatBoost
model_laps_catboost = CatBoostRegressor(iterations= 3500,learning_rate= 0.08,depth= 9,objective='RMSE',
            eval_metric= 'RMSE',verbose= 500,border_count=254,task_type='CPU',use_best_model=False)
model_laps_catboost.fit(X_train_laps, y_train_laps, cat_features=cat_cols)
y_pred_laps_catboost = model_laps_catboost.predict(X_test_laps)

laps_mse_catboost = mean_squared_error(y_test_laps, y_pred_laps_catboost)
laps_mse_catboost

# Model 4: Predicting Avg_Lap_Time (Regression) - XGBoost
model_avg_lap_time_xgb = XGBRegressor(n_estimators= 3500,learning_rate= 0.05,max_depth= 5, objective='reg:squarederror', 
                              eval_metric= 'rmse', enable_categorical=True, random_state=42)
model_avg_lap_time_xgb.fit(X_train_avg_lap_time, y_train_avg_lap_time)
y_pred_avg_lap_time_xgb = model_avg_lap_time_xgb.predict(X_test_avg_lap_time)

avg_lap_time_mse_xgb = mean_squared_error(y_test_avg_lap_time, y_pred_avg_lap_time_xgb)
avg_lap_time_mse_xgb

# Model 4.2: Predicting Avg_Lap_Time (Regression) - LightGBM
model_avg_lap_time_lgbm = lgb.LGBMRegressor(n_estimators= 3500,learning_rate= 0.08,max_depth= 9, objective='regression', metric='rmse',
                              random_state=42, verbose=-1, num_leaves=45, subsample=0.6, colsample_bytree=0.9)
model_avg_lap_time_lgbm.fit(X_train_avg_lap_time, y_train_avg_lap_time)
y_pred_avg_lap_time_lgbm = model_avg_lap_time_lgbm.predict(X_test_avg_lap_time)

avg_lap_time_mse_lgbm = mean_squared_error(y_test_avg_lap_time, y_pred_avg_lap_time_lgbm)
avg_lap_time_mse_lgbm

# Model 4.3: Predicting Avg_Lap_Time (Regression) - CatBoost
model_avg_lap_time_catboost = CatBoostRegressor(iterations= 3500,learning_rate= 0.08,depth= 9,objective='RMSE',
            eval_metric= 'RMSE',verbose= 500,border_count=254,task_type='CPU',use_best_model=False)
model_avg_lap_time_catboost.fit(X_train_avg_lap_time, y_train_avg_lap_time, cat_features=cat_cols)
y_pred_avg_lap_time_catboost = model_avg_lap_time_catboost.predict(X_test_avg_lap_time)

avg_lap_time_mse_catboost = mean_squared_error(y_test_avg_lap_time, y_pred_avg_lap_time_catboost)
avg_lap_time_mse_catboost

# Saving the models
import joblib

joblib.dump(model_stints_xgb, 'model_stints.pkl')
joblib.dump(model_tires_lgbm, 'model_tires.pkl')
joblib.dump(model_laps_xgb, 'model_laps.pkl')
joblib.dump(model_avg_lap_time_xgb, 'model_avg_lap_time.pkl')

# load the models 
loaded_model_stints = joblib.load('model_stints.pkl')
loaded_model_tires = joblib.load('model_tires.pkl')
loaded_model_laps = joblib.load('model_laps.pkl')
loaded_model_avg_lap_time = joblib.load('model_avg_lap_time.pkl')

data_cleaned['EventName'].unique(), data_cleaned['Driver'].unique(), data_cleaned['GridPosition'].unique(), data_cleaned['Year'].unique()

"""
Current Formula 1 Drivers (2024 season):
VER – Max Verstappen
PER – Sergio Pérez
SAI – Carlos Sainz
LEC – Charles Leclerc
RUS – George Russell
NOR – Lando Norris
HAM – Lewis Hamilton
PIA – Oscar Piastri
ALO – Fernando Alonso
STR – Lance Stroll
ZHO – Zhou Guanyu
MAG – Kevin Magnussen
RIC – Daniel Ricciardo
TSU – Yuki Tsunoda
ALB – Alex Albon
HUL – Nico Hülkenberg
OCO – Esteban Ocon
GAS – Pierre Gasly
BOT – Valtteri Bottas
SAR – Logan Sargeant
"""

data_cleaned['TyreLife'].unique(), data_cleaned['FreshTyre'].unique(), data_cleaned['TrackStatus'].unique(), data_cleaned['Rainfall'].unique()

data_cleaned['Speed'].unique(), data_cleaned['Throttle'].unique(), data_cleaned['Brake'].unique(), data_cleaned['DRS'].unique()

data_cleaned['ClassifiedPosition'].unique(), data_cleaned['AirTemp'].unique()

data_cleaned['Humidity'].unique(), data_cleaned['Pressure'].unique(), data_cleaned['TrackTemp'].unique(), data_cleaned['WindSpeed'].unique()

# Test inputs to make predictions (example inputs)
test_input = pd.DataFrame({
    'EventName': ['Mexican Grand Prix'],     # Categorical - event name
    'Driver': ['NOR'],                       # Categorical - driver name
    'GridPosition': [10],                     # Integer - starting position
    'Year': [2024],                          # Categorical - race year
    'TyreLife': [10],                         # Integer - tire age in laps
    'FreshTyre': ['False'],                   # Boolean - fresh tires (True means fresh)
    'TrackStatus': [2671],                      # Categorical - track status (1: clear)
    'Rainfall': ['True'],                   # Boolean - rainfall (True means rainy)
    'Speed': [285.0],                        # Float - car speed in km/h
    'Throttle': [85],                        # Integer - throttle percentage
    'Brake': ['False'],                       # Boolean - brake applied (False: no)
    'DRS': [1],                              # Categorical - DRS active (1: yes)
    'ClassifiedPosition': [1],               # Categorical - current race position
    'AirTemp': [30],                       # Float - air temperature in Celsius
    'Humidity': [80.0],                      # Float - humidity percentage
    'Pressure': [990.0],                    # Float - air pressure in hPa
    'TrackTemp': [35],                     # Float - track temperature in Celsius
    'WindSpeed': [1]                       # Float - wind speed in km/h
})

# Convert categorical columns to 'category' dtype
test_input['Driver'] = test_input['Driver'].astype('category')
test_input['EventName'] = test_input['EventName'].astype('category')
test_input['ClassifiedPosition'] = test_input['ClassifiedPosition'].astype('category')
test_input['FreshTyre'] = test_input['FreshTyre'].astype('category')
test_input['Brake'] = test_input['Brake'].astype('category')
test_input['Rainfall'] = test_input['Rainfall'].astype('category')
test_input['Year'] = test_input['Year'].astype('category')
test_input['TrackStatus'] = test_input['TrackStatus'].astype('category')
test_input['DRS'] = test_input['DRS'].astype('category')

# Predictions from loaded models
# Model 1 - Number of Stints
pred_stints = loaded_model_stints.predict(test_input)
pred_stints = round(pred_stints[0])  # Rounding the prediction

print(f'Predicted Number of Stints: {pred_stints}')

# (pred_stints, pred_tires, pred_laps, pred_avg_lap_time)

# Model 2- Used Tire Compound
pred_tires = loaded_model_tires.predict(test_input)
print(f'Predicted Tire Compounds Used: {pred_tires}')

# Model 3 - Lap Number per Stint
pred_laps = loaded_model_laps.predict(test_input)
pred_laps = round(pred_laps[0])  # Rounding the prediction

print(f'Predicted Number of Laps within Each Stint: {pred_laps}')

# Model 4 - Average Lap Time per Stint
pred_avg_lap_time = loaded_model_avg_lap_time.predict(test_input)
print(f'Predicted Average Lap Time per Stint: {pred_avg_lap_time}')
